refactor: split codegen.rs into a module and dedicated emit_expr/emit_stmt files
